customizations/pyproject.toml
[project]
name = "Customizations"
version = "0.1.0"
description = "Customizations for KeyGen-ESET"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.11.18",
    "python-dotenv==1.1.0",
    "python-telegram-bot==22.1"
]
